Refactored getPremium logic to use PremiumCalculator module
This commit moves the logic for calculating the premium in MeowMed>EDA>PolicyService to a new module called PremiumCalculator unter MeowMed>EDA>PolicyService>Module> . All the relevant methods for calculating the premium have been implemented and properly documented. This helps to improve the code organization and makes it easier to maintain and update the premium calculation logic in the future.